Fixed Small Sound Bug
I wrote a midi song and am wondering if I can get it to play on repeat
in the background or if matlab will even be able to import the files. If
you can read them, take a listen. This is only my first experience with
midi, but I think I did a pretty good job.

BTW, the midi2audio.m and readmidi.m are copyright protected by some
other dude and are not mine...