Final Matrix Display
Previously, when you lost the game, it would just say "Game Over." Now
it will display the matrix so you can see exactly how you lost. I should
make sure that it displays the matrix when you win and give the player
the option of continuing on. Oh! A score counter of some kind would be
pretty awesome too!